refactor(plugins): apply /simplify review — reuse, dedup, dead code
Quality cleanups from the /simplify pass on slice 5a (no behaviour change
except detector process matching now matches the app-pause guard):

- Reuse: extract run_loop's process_match_lower into a shared crate::proc_match
  module; detector process_running now uses the same token-aware matcher, so
  `zoom` matches Zoom but not `zoominfo` (was naive substring). Tests move with
  the function; run_loop imports it.
- Reuse: read_flag now uses secure_io::read_capped instead of hand-rolling
  metadata + read_to_string.
- Dead code: drop host_function_names_for (build_sandboxed_plugin inlines its
  own dedup); remove its test and re-export.
- Simplification: collapse the two near-identical register_capability arms into
  a generic register_probe helper.

Skipped (noted, not applied): the altitude suggestion to move detector params
into the Capability enum — process_name is a config param, not a permission
scope, so the manifest is its right home; refactoring is speculative until a
second param appears. Micro-efficiency notes (module.to_vec, per-process
lowercase) are one-time/throttled and left as-is.

/security-review found no vulnerabilities in the slice.

detect.rs/proc_match.rs at 100% line coverage; full Rust lib suite 982 green;
clippy/fmt/cargo-doc clean.
